refactor: #90 어쿠스틱&클래식 기타 package 분리
