+++ src/main/java/com/ajou/hertz/domain/instrument/acoustic_and_classic_guitar/constant/AcousticAndClassicGuitarBrand.java
@@ -1,6 +1,8 @@
-package com.ajou.hertz.domain.instrument.constant;
+package com.ajou.hertz.domain.instrument.acoustic_and_classic_guitar.constant;
 
 import static com.ajou.hertz.domain.instrument.constant.ProductOrigin.*;
+
+import com.ajou.hertz.domain.instrument.constant.ProductOrigin;
 
 import lombok.AllArgsConstructor;
