+++ src/main/java/com/ajou/hertz/domain/instrument/acoustic_and_classic_guitar/constant/AcousticAndClassicGuitarModel.java
@@ -1,4 +1,4 @@
-package com.ajou.hertz.domain.instrument.constant;
+package com.ajou.hertz.domain.instrument.acoustic_and_classic_guitar.constant;
 
 public enum AcousticAndClassicGuitarModel {
