+++ src/main/java/com/ajou/hertz/domain/instrument/acoustic_and_classic_guitar/dto/response/AcousticAndClassicGuitarResponse.java
@@ -1,15 +1,17 @@
-package com.ajou.hertz.domain.instrument.dto.response;
+package com.ajou.hertz.domain.instrument.acoustic_and_classic_guitar.dto.response;
 
 import java.util.List;
 
 import com.ajou.hertz.common.dto.response.AddressResponse;
-import com.ajou.hertz.domain.instrument.constant.AcousticAndClassicGuitarBrand;
-import com.ajou.hertz.domain.instrument.constant.AcousticAndClassicGuitarModel;
-import com.ajou.hertz.domain.instrument.constant.AcousticAndClassicGuitarPickUp;
-import com.ajou.hertz.domain.instrument.constant.AcousticAndClassicGuitarWood;
+import com.ajou.hertz.domain.instrument.acoustic_and_classic_guitar.constant.AcousticAndClassicGuitarBrand;
+import com.ajou.hertz.domain.instrument.acoustic_and_classic_guitar.constant.AcousticAndClassicGuitarModel;
+import com.ajou.hertz.domain.instrument.acoustic_and_classic_guitar.constant.AcousticAndClassicGuitarPickUp;
+import com.ajou.hertz.domain.instrument.acoustic_and_classic_guitar.constant.AcousticAndClassicGuitarWood;
 import com.ajou.hertz.domain.instrument.constant.InstrumentCategory;
 import com.ajou.hertz.domain.instrument.constant.InstrumentProgressStatus;
-import com.ajou.hertz.domain.instrument.dto.AcousticAndClassicGuitarDto;
+import com.ajou.hertz.domain.instrument.acoustic_and_classic_guitar.dto.AcousticAndClassicGuitarDto;
+import com.ajou.hertz.domain.instrument.dto.response.InstrumentImageResponse;
+import com.ajou.hertz.domain.instrument.dto.response.InstrumentResponse;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.AccessLevel;
